requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pdfplumber
python-docx
openai
gtts
python-dotenv
huggingface_hub
transformers
torch
soundfile
requests
datasets
scipy
